IOS-42 feat: FortuneFront, Back으로 view 분리 후 적용
